+++ src/com/ACM/binarycalculator/CalculatorOctalFragment.java
@@ -1,5 +1,7 @@
 package com.ACM.binarycalculator;
 
+import java.math.BigInteger;
+
 import android.app.Activity;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -351,8 +353,9 @@ public class CalculatorOctalFragment extends Fragment {
 	public void updateWorkingTextView(String dataToBePassed, int base) {
 		if (dataToBePassed.length() != 0) {
 
-			mCurrentWorkingText = Integer.toOctalString(Integer.parseInt(
+			mCurrentWorkingText = Long.toOctalString(Long.parseLong(
 					dataToBePassed, base));
+
 			mWorkingTextView.setText(mCurrentWorkingText);
 
 		} else {
